Move navbar into SinglePageApp
