+++ frontend/src/components/SinglePageApp/Navbar/Navbar.tsx
