+++ frontend/src/components/SinglePageApp/Navbar/UserInfoName/index.ts
