+++ frontend/src/components/SinglePageApp/Navbar/UserPicture/index.ts
