+++ frontend/src/components/SinglePageApp/Navbar/index.ts
